# api/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benevolus",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/campaigns/{campaign}/donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "List campaign donations",
                "operationId": "listCampaignDonations",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignDonationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/donations": {
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Create a new donation",
                "operationId": "createDonation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DonationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donation": {
                                            "$ref": "#/components/schemas/DonationResource"
                                        },
                                        "payment": {
                                            "$ref": "#/components/schemas/PaymentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/donations/{id}/confirm": {
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Confirm donation payment",
                "operationId": "confirmDonation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "External reference ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Donation confirmed successfully"
                    }
                }
            }
        },
...
        "/api/campaigns/{campaign}/reports": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create report",
                "operationId": "createReport",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user profile",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Profile updated successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DonorRequest": {
                "required": [
                    "name",
                    "email",
                    "taxId",
                    "phoneNumber"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "taxId": {
                        "type": "string",
                        "pattern": "^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$"
                    },
                    "phoneNumber": {
                        "type": "string",
                        "pattern": "^\\d{10,11}$"
                    }
                },
                "type": "object"
            },
            "DonationRequest": {
                "required": [
                    "amount",
                    "anonymousDonation",
                    "donor",
                    "paymentMethod"
                ],
                "properties": {
                    "amount": {
                        "type": "number",
                        "maximum": 1000000,
                        "minimum": 1
                    },
                    "anonymousDonation": {
                        "type": "boolean"
                    },
                    "campaignId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "donor": {
                        "$ref": "#/components/schemas/DonorRequest"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "pix",
                            "credit_card",
                            "boleto"
                        ]
                    }
                },
                "type": "object"
            },
            "StoreReportRequest": {
                "required": [
                    "type",
                    "reason",
                    "description"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "personal_rights",
                            "terms_violations"
                        ]
                    },
                    "reason": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 10
                    }
                },
                "type": "object"
            },
            "UpdateProfileRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "number": {
                                "type": "string",
                                "maxLength": 20
                            },
                            "city": {
                                "type": "string",
                                "maxLength": 100
                            },
                            "state": {
                                "type": "string",
                                "maxLength": 50
                            },
                            "zipcode": {
                                "type": "string",
                                "maxLength": 20
                            },
                            "country": {
                                "type": "string",
                                "maxLength": 100
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "CampaignDonationResource": {
                "properties": {
                    "donor_name": {
                        "type": "string"
                    },
                    "amount_cents": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "CampaignResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "goal": {
                        "type": "number"
                    },
                    "raised": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "DonationResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "amountCents": {
                        "type": "number"
                    },
                    "isAnonymous": {
                        "type": "boolean"
                    },
                    "paymentStatus": {
                        "type": "string"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "externalReference": {
                        "type": "string",
                        "nullable": true
                    },
                    "paidAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "donor": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DonorResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "campaign": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CampaignResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "PaymentResource": {
                "properties": {
                    "method": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "pixCode": {
                        "type": "string"
                    },
                    "qrCodeUrl": {
                        "type": "string"
                    },
                    "expiresAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ReportResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "campaign": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CampaignResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "DonorResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string",
                                "nullable": true
                            },
                            "number": {
                                "type": "string",
                                "nullable": true
                            },
                            "city": {
                                "type": "string",
                                "nullable": true
                            },
                            "state": {
                                "type": "string",
                                "nullable": true
                            },
                            "zipcode": {
                                "type": "string",
                                "nullable": true
                            },
                            "country": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}